Proper Proper Init
